+++ src/functionalTest/groovy/com/autonomousapps/android/projects/NativeLibProject.groovy
@@ -1,0 +1,55 @@
+package com.autonomousapps.android.projects
+
+import com.autonomousapps.AbstractProject
+import com.autonomousapps.AdviceHelper
+import com.autonomousapps.advice.Advice
+import com.autonomousapps.kit.*
+
+import static com.autonomousapps.AdviceHelper.dependency
+import static com.autonomousapps.AdviceHelper.transitiveDependency
+
+final class NativeLibProject extends AbstractProject {
+
+  final GradleProject gradleProject
+  private final String agpVersion
+
+  NativeLibProject(String agpVersion) {
+    this.agpVersion = agpVersion
+    this.gradleProject = build()
+  }
+
+  private GradleProject build() {
+    def builder = newGradleProjectBuilder()
+    builder.withRootProject { root ->
+      root.gradleProperties = GradleProperties.minimalAndroidProperties()
+      root.withBuildScript { bs ->
+        bs.buildscript = BuildscriptBlock.defaultAndroidBuildscriptBlock(agpVersion)
+      }
+    }
+    builder.withAndroidSubproject('app') { a ->
+      a.withBuildScript { bs ->
+        bs.plugins = [Plugin.androidAppPlugin, Plugin.kotlinAndroidPlugin]
+        bs.repositories = [Repository.LIBS]
+        bs.android = AndroidBlock.defaultAndroidAppBlock(true)
+        bs.dependencies = dependencies
+      }
+    }
+
+    def project = builder.build()
+    project.writer().write()
+    return project
+  }
+
+  private List<Dependency> dependencies = [
+    Dependency.kotlinStdLib("implementation"),
+    Dependency.appcompat("implementation"),
+    Dependency.constraintLayout("implementation"),
+    new Dependency("implementation", "amazon-chime-sdk-media", "aar")
+  ]
+
+  List<Advice> actualAdvice() {
+    return AdviceHelper.actualAdviceForFirstSubproject(gradleProject)
+  }
+
+  final List<Advice> expectedAdvice = []
+}